% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilez.R
\name{add_up}
\alias{add_up}
\title{Add simulated outputs across objects}
\usage{
add_up(G)
}
\arguments{
\item{G}{a list of \code{\link{gilez}} objects}
}
\value{
a \code{\link{gilez}} object
}
\description{
Add simulated outputs across objects
}
