% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilers.R
\name{among}
\alias{among}
\title{A wrapper for \code{\%in\%}}
\usage{
among(x, y = 0, ...)
}
\arguments{
\item{x}{the input.}

\item{y}{value(s) of \code{x} corresponding to success. Default \code{0}.}

\item{...}{ignored.}
}
\value{
an object the same shape as \code{x} and \code{mode} \code{logical}.
}
\description{
A wrapper for \code{\%in\%}
}
