% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.R
\name{bestz}
\alias{bestz}
\title{Pick out best levels of each X variable at which to view effects.}
\usage{
bestz(m, x)
}
\arguments{
\item{m}{a model.}

\item{x}{a population.}
}
\value{
a tall data.frame with columns \code{term}, \code{level}, and \code{value}.
}
\description{
This function kind of assumes the goal is \code{\link{gdiff}}.
}
