% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boil.R
\name{boil}
\alias{boil}
\title{Take the weighted mean of a function of a vector}
\usage{
boil(x, f = value, w = rep(1, length(x)), ...)
}
\arguments{
\item{x}{the input.}

\item{f}{a \code{function}, typically to convert \code{x} to \code{logical}. Default \code{\link{value}} (atypical).}

\item{w}{weights. Default \code{1}.}

\item{...}{other arguments to \code{f}. Most cases take an argument \code{y} of values corresponding to success.}
}
\value{
a data.frame with a numeric value for each possible outcome.
}
\description{
Take the weighted mean of a function of a vector
}
