% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consider.R, R/crch.R, R/svymnlogit.R
\name{consider}
\alias{consider}
\alias{consider.default}
\alias{consider.crch}
\alias{consider.svymnlogit}
\title{Draw \code{n} _parameter_ vectors from the fitted likelihood for a model}
\usage{
consider(m, x, n = 1, ...)

\method{consider}{default}(m, x, n = 1, ...)

\method{consider}{crch}(m, x, n = 1, ...)

\method{consider}{svymnlogit}(m, x, n = 1, ...)
}
\arguments{
\item{m}{the model.}

\item{x}{data.frame. The population of interest.}

\item{n}{numeric. Number of vectors to draw. Default \code{1}.}

\item{...}{other arguments to particular methods, I guess}
}
\value{
a matrix of parameter vectors.
}
\description{
Draw \code{n} _parameter_ vectors from the fitted likelihood for a model
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for model objects not needing special handling.

\item \code{crch}: crch

\item \code{svymnlogit}: svymnlogit
}}

