% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R, R/svyolr.r
\name{draw}
\alias{draw}
\alias{draw.svyolr}
\title{Draw \code{n} response vectors from a fitted model for a population}
\usage{
draw(m, x, n = 1, ...)

\method{draw}{svyolr}(m, x, n = 1, ...)
}
\arguments{
\item{m}{the model.}

\item{x}{data.frame. The population of interest.}

\item{n}{numeric. Number of vectors to draw. Default \code{1}.}

\item{...}{other arguments to \code{weights} or to methods.}
}
\value{
a vector of responses.
}
\description{
Draw \code{n} response vectors from a fitted model for a population
}
\section{Methods (by class)}{
\itemize{
\item \code{svyolr}: the \code{x} argument may not be omitted
}}

