% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilers.R
\name{equal}
\alias{equal}
\title{A wrapper for \code{among}, confusingly}
\usage{
equal(x, y = unique(x), ...)
}
\arguments{
\item{x}{the input.}

\item{y}{value(s) of \code{x} corresponding to success. Default \code{unique(x)}.}

\item{...}{ignored.}
}
\value{
a list of results for \code{among}, one for each separate value in \code{y}
}
\description{
A wrapper for \code{among}, confusingly
}
