% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilez.R
\name{gdiff}
\alias{gdiff}
\title{Contrast effects for different levels with each other}
\usage{
gdiff(Y)
}
\arguments{
\item{Y}{a \code{\link{gilez}} object}
}
\value{
another \code{gilez} object but with more columns
}
\description{
Contrast effects for different levels with each other
}
