% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweights.R, R/svyolr.r
\name{getweights}
\alias{getweights}
\alias{getweights.default}
\alias{getweights.svyolr}
\title{Get weights from a model object}
\usage{
getweights(m, ...)

\method{getweights}{default}(m, ...)

\method{getweights}{svyolr}(m, x, ...)
}
\arguments{
\item{m}{the model object.}

\item{...}{other arguments to \code{weights} or to methods.}

\item{x}{the population.}
}
\value{
a vector of weights.
}
\description{
Get weights from a model object
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method

\item \code{svyolr}: svyolr
}}

