% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilez.R
\name{gilez}
\alias{gilez}
\title{Substantive effects via simulation}
\usage{
gilez(m, x = stats::model.frame(m), Z = pick(m), f = value, y = 0,
  n = 1000, ...)
}
\arguments{
\item{m}{a model.}

\item{x}{a population within which to simulate effects. Default \code{model.frame(m)} (the original population).}

\item{Z}{a named ist of two or more contrasting values for each variable. Default \code{pick(m)}.}

\item{f}{a function for boiling down results. This is a function, not a string. Default \code{value}.}

\item{y}{values of the outcome representing success under \code{f}. Default \code{0}.}

\item{n}{number of simulated response vectors to generate for each level for each test. Default \code{1000}.}

\item{...}{other arguments to functions used within.}
}
\value{
a \code{sumer} object (a tidy summary, with model metadata as an attribute). (Well, eventually.)
}
\description{
Substantive effects via simulation
}
