% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilez.R
\name{pick}
\alias{pick}
\title{Functional stub for smart contrasting-value function}
\usage{
pick(m)
}
\arguments{
\item{m}{a model}
}
\value{
for now, a list of numeric vectors, each with value \code{c(0, 1)}, one for each predictor in the model
}
\description{
Functional stub for smart contrasting-value function
}
