% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svymnlogit.R, R/termz.R
\name{termz.svymnlogit}
\alias{termz.svymnlogit}
\alias{termz}
\alias{termz.default}
\title{Wrapper for best way to pull out model terms}
\usage{
\method{termz}{svymnlogit}(m)

termz(m)

\method{termz}{default}(m)
}
\arguments{
\item{m}{the model.}
}
\value{
a \code{terms} object
}
\description{
Write a method for your model \code{m} if \code{stats::\link[stats]{terms}(m)} does not work
}
\section{Methods (by class)}{
\itemize{
\item \code{svymnlogit}: svymnlogit has no \code{terms} element

\item \code{default}: For most models, just a wrapper for \code{stats::terms}
}}

