% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilers.R
\name{value}
\alias{value}
\title{Output your input (the trivial boiler)}
\usage{
value(x, ...)
}
\arguments{
\item{x}{the output.}

\item{...}{ignored.}
}
\value{
the input.
}
\description{
Output your input (the trivial boiler)
}
