% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{weighted.sum}
\alias{weighted.sum}
\title{Weighted sum g-function
Just \code{sum(x * w)} in a nice wrapper}
\usage{
weighted.sum(x, w = 1, ...)
}
\arguments{
\item{x}{boiler output}

\item{w}{"weights" (which should really be counts). Default \code{1}.}

\item{...}{ignored.}
}
\value{
\code{sum(x * w)}
}
\description{
Weighted sum g-function
Just \code{sum(x * w)} in a nice wrapper
}
