% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betareg.R, R/crch.R, R/draw.R, R/glm.R,
%   R/svymnlogit.R, R/svyolr.R
\name{draw.betareg}
\alias{draw.betareg}
\alias{draw.crch}
\alias{draw}
\alias{draw.glm}
\alias{draw.svymnlogit}
\alias{draw.svyolr}
\title{Draw \code{n} response vectors from a fitted model for a population}
\usage{
\method{draw}{betareg}(m, x = stats::model.frame(m), n = 1, ...)

\method{draw}{crch}(m, x = stats::model.frame(m), n = 1, ...)

draw(m, x, n = 1, ...)

\method{draw}{glm}(m, x = stats::model.frame(m), n = 1, ...)

\method{draw}{svymnlogit}(m, x = stats::model.frame(m), n = 1, ...)

\method{draw}{svyolr}(m, x, n = 1, ...)
}
\arguments{
\item{m}{the model.}

\item{x}{data.frame. The population of interest.}

\item{n}{numeric. Number of vectors to draw. Default \code{1}.}

\item{...}{other arguments to \code{weights} or to methods.}
}
\value{
a vector of responses.
}
\description{
Draw \code{n} response vectors from a fitted model for a population
}
\section{Methods (by class)}{
\itemize{
\item \code{betareg}: betareg

\item \code{crch}: crch

\item \code{glm}: glm

\item \code{svymnlogit}: svymnlogit

\item \code{svyolr}: the \code{x} argument may not be omitted
}}

