% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betareg.R, R/crch.R, R/getweights.R, R/glm.R,
%   R/svymnlogit.R, R/svyolr.R
\name{getweights.betareg}
\alias{getweights.betareg}
\alias{getweights.crch}
\alias{getweights}
\alias{getweights.default}
\alias{getweights.glm}
\alias{getweights.svymnlogit}
\alias{getweights.svyolr}
\title{Get weights from a model object}
\usage{
\method{getweights}{betareg}(m, x, ...)

\method{getweights}{crch}(m, x, ...)

getweights(m, ...)

\method{getweights}{default}(m, ...)

\method{getweights}{glm}(m, x, ...)

\method{getweights}{svymnlogit}(m, ...)

\method{getweights}{svyolr}(m, x, ...)
}
\arguments{
\item{m}{the model object.}

\item{x}{the population}

\item{...}{other arguments to \code{weights} or to methods.}
}
\value{
a vector of weights.
}
\description{
Get weights from a model object
}
\section{Methods (by class)}{
\itemize{
\item \code{betareg}: betareg

\item \code{crch}: crch

\item \code{default}: default method

\item \code{glm}: glm

\item \code{svymnlogit}: svymnlogit

\item \code{svyolr}: svyolr
}}

