% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilez.R
\name{gilez}
\alias{gilez}
\title{Substantive effects via simulation}
\usage{
gilez(m, x = stats::model.frame(m), Z = bestz(m, x), f = value,
  y = 0, n = 1000, parallel = FALSE, w = getweights(m, x),
  g = stats::weighted.mean, ...)
}
\arguments{
\item{m}{a model.}

\item{x}{a population within which to simulate effects. Default \code{model.frame(m)} (the original population).}

\item{Z}{a named ist of two or more contrasting values for each variable. Default \code{bestz(m, x)}.}

\item{f}{a function for boiling down results. This is a function, not a string. Default \code{value}.}

\item{y}{values of the outcome representing success under \code{f}. Default \code{0}.}

\item{n}{number of simulated response vectors to generate for each level for each test. Default \code{1000}.}

\item{parallel}{logical. Passed on to \code{\link[plyr]{ddply}} at the outermost (term and level) levels.}

\item{w}{weights with respect to which to take averages. Default \code{\link{getweights}(m, x)}.}

\item{g}{function. the desired one-number summary of the boiled values. Default \code{stats::\link[stats]{weighted.mean}}.}

\item{...}{other arguments to functions used within.}
}
\value{
a tall \code{data.frame} with class \code{gilez} to pass to \code{\link{gdiff}} or \code{\link[wickr]{sumer}}.
}
\description{
Substantive effects via simulation
}
