% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagine.R
\name{imagine}
\alias{imagine}
\title{Combine draw, getweights, and boil}
\usage{
imagine(z, m, x, f, y, n)
}
\arguments{
\item{z}{a row of a data.frame given by \code{pick}.}

\item{m}{a model.}

\item{x}{a data set; a population.}

\item{f}{a boiler.}

\item{y}{values of the dependent variable representing success under \code{f}.}

\item{n}{number of response vectors to simulate.}
}
\value{
a \code{data.frame} of simulated outcome summary statistics.
}
\description{
This is exported, but don't call it directly.
}
