% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{pick}
\alias{pick.factor}
\alias{pick.logical}
\alias{pick.character}
\alias{pick.integer}
\alias{pick.numeric}
\title{Pick out best levels of a given X variable at which to view effects.}
\usage{
pick(x)

\method{pick}{factor}(x)

\method{pick}{logical}(x)

\method{pick}{character}(x)

\method{pick}{integer}(x)

\method{pick}{numeric}(x)
}
\arguments{
\item{x}{a column from a population.}
}
\value{
a tall data.frame with columns \code{level} and \code{value}.
}
\description{
Pick out best levels of a given X variable at which to view effects.
}
\section{Methods (by class)}{
\itemize{
\item \code{factor}: return all levels of the factor

\item \code{logical}: return FALSE and TRUE

\item \code{character}: return all unique elements, sorted

\item \code{integer}: return all unique elements, sorted

\item \code{numeric}: return the levels in the data and a one-unit increase
}}

